Added assert tests to the Azure Login test (test title and login button)
https://stackoverflow.com/questions/74040881/how-to-use-expect-in-playwright-with-c-sharp-xunit
